chore: refactor lessons into separate modules
