+++ slides/src/Slides/Layouts/index.js
@@ -1,0 +1,4 @@
+import Lesson from "./Lesson.js";
+import TitleSlide from "./TitleSlide.js";
+
+export { Lesson, TitleSlide };
