+++ slides/src/Slides/Lessons/Lesson02.js
@@ -1,0 +1,73 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson02() {
+    return html`
+        <${Lesson} title="Exploring the Terminal" lessonId="lesson_02" subtitle="Lesson 2">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide43.png" alt="Slide 43" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide44.png" alt="Slide 44" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide45.png" alt="Slide 45" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide46.png" alt="Slide 46" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide47.png" alt="Slide 47" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide48.png" alt="Slide 48" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide49.png" alt="Slide 49" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide50.png" alt="Slide 50" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide51.png" alt="Slide 51" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide52.png" alt="Slide 52" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide53.png" alt="Slide 53" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide54.png" alt="Slide 54" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide55.png" alt="Slide 55" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide56.png" alt="Slide 56" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide57.png" alt="Slide 57" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide58.png" alt="Slide 58" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide59.png" alt="Slide 59" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide60.png" alt="Slide 60" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide61.png" alt="Slide 61" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide62.png" alt="Slide 62" />
+            </section>
+        <//>`;
+}
+
+export default Lesson02;
