+++ slides/src/Slides/Lessons/Lesson03.js
@@ -1,0 +1,73 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson03() {
+    return html`
+        <${Lesson} title="How Computers Work" lessonId="lesson_03" subtitle="Lesson 3">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide64.png" alt="Slide 64" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide65.png" alt="Slide 65" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide66.png" alt="Slide 66" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide67.png" alt="Slide 67" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide68.png" alt="Slide 68" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide69.png" alt="Slide 69" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide70.png" alt="Slide 70" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide71.png" alt="Slide 71" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide72.png" alt="Slide 72" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide73.png" alt="Slide 73" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide74.png" alt="Slide 74" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide75.png" alt="Slide 75" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide76.png" alt="Slide 76" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide77.png" alt="Slide 77" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide78.png" alt="Slide 78" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide79.png" alt="Slide 79" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide80.png" alt="Slide 80" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide81.png" alt="Slide 81" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide82.png" alt="Slide 82" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide83.png" alt="Slide 83" />
+            </section>
+        <//>`;
+}
+
+export default Lesson03;
