+++ slides/src/Slides/Lessons/Lesson04.js
@@ -1,0 +1,79 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson04() {
+    return html`
+        <${Lesson} title="How to be a Great SWE" lessonId="lesson_04" subtitle="Lesson 4">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide85.png" alt="Slide 85" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide86.png" alt="Slide 86" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide87.png" alt="Slide 87" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide88.png" alt="Slide 88" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide89.png" alt="Slide 89" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide90.png" alt="Slide 90" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide91.png" alt="Slide 91" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide92.png" alt="Slide 92" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide93.png" alt="Slide 93" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide94.png" alt="Slide 94" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide95.png" alt="Slide 95" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide96.png" alt="Slide 96" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide97.png" alt="Slide 97" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide98.png" alt="Slide 98" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide99.png" alt="Slide 99" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide100.png" alt="Slide 100" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide101.png" alt="Slide 101" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide102.png" alt="Slide 102" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide103.png" alt="Slide 103" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide104.png" alt="Slide 104" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide105.png" alt="Slide 105" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide106.png" alt="Slide 106" />
+            </section>
+        <//>`;
+}
+
+export default Lesson04;
