+++ slides/src/Slides/Lessons/Lesson05.js
@@ -1,0 +1,112 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson05() {
+    return html`
+        <${Lesson} title="Software Dev Life Cycle" lessonId="lesson_05" subtitle="Lesson 5">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide108.png" alt="Slide 108" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide109.png" alt="Slide 109" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide110.png" alt="Slide 110" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide111.png" alt="Slide 111" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide112.png" alt="Slide 112" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide113.png" alt="Slide 113" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide114.png" alt="Slide 114" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide115.png" alt="Slide 115" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide116.png" alt="Slide 116" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide117.png" alt="Slide 117" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide118.png" alt="Slide 118" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide119.png" alt="Slide 119" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide120.png" alt="Slide 120" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide121.png" alt="Slide 121" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide122.png" alt="Slide 122" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide123.png" alt="Slide 123" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide124.png" alt="Slide 124" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide125.png" alt="Slide 125" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide126.png" alt="Slide 126" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide127.png" alt="Slide 127" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide128.png" alt="Slide 128" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide129.png" alt="Slide 129" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide130.png" alt="Slide 130" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide131.png" alt="Slide 131" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide132.png" alt="Slide 132" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide133.png" alt="Slide 133" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide134.png" alt="Slide 134" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide135.png" alt="Slide 135" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide136.png" alt="Slide 136" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide137.png" alt="Slide 137" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide138.png" alt="Slide 138" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide139.png" alt="Slide 139" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide140.png" alt="Slide 140" />
+            </section>
+        <//>`;
+}
+
+export default Lesson05;
