+++ slides/src/Slides/Lessons/Lesson06.js
@@ -1,0 +1,76 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson06() {
+    return html`
+        <${Lesson} title="Statements and Variables" lessonId="lesson_06" subtitle="Lesson 6">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide142.png" alt="Slide 142" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide143.png" alt="Slide 143" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide144.png" alt="Slide 144" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide145.png" alt="Slide 145" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide146.png" alt="Slide 146" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide147.png" alt="Slide 147" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide148.png" alt="Slide 148" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide149.png" alt="Slide 149" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide150.png" alt="Slide 150" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide151.png" alt="Slide 151" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide152.png" alt="Slide 152" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide153.png" alt="Slide 153" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide154.png" alt="Slide 154" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide155.png" alt="Slide 155" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide156.png" alt="Slide 156" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide157.png" alt="Slide 157" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide158.png" alt="Slide 158" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide159.png" alt="Slide 159" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide160.png" alt="Slide 160" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide161.png" alt="Slide 161" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide162.png" alt="Slide 162" />
+            </section>
+        <//>`;
+}
+
+export default Lesson06;
