+++ slides/src/Slides/Lessons/Lesson07.js
@@ -1,0 +1,139 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson07() {
+    return html`
+        <${Lesson} title="Control Flow" lessonId="lesson_07" subtitle="Lesson 7">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide164.png" alt="Slide 164" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide165.png" alt="Slide 165" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide166.png" alt="Slide 166" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide167.png" alt="Slide 167" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide168.png" alt="Slide 168" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide169.png" alt="Slide 169" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide170.png" alt="Slide 170" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide171.png" alt="Slide 171" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide172.png" alt="Slide 172" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide173.png" alt="Slide 173" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide174.png" alt="Slide 174" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide175.png" alt="Slide 175" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide176.png" alt="Slide 176" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide177.png" alt="Slide 177" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide178.png" alt="Slide 178" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide179.png" alt="Slide 179" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide180.png" alt="Slide 180" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide181.png" alt="Slide 181" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide182.png" alt="Slide 182" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide183.png" alt="Slide 183" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide184.png" alt="Slide 184" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide185.png" alt="Slide 185" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide186.png" alt="Slide 186" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide187.png" alt="Slide 187" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide188.png" alt="Slide 188" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide189.png" alt="Slide 189" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide190.png" alt="Slide 190" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide191.png" alt="Slide 191" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide192.png" alt="Slide 192" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide193.png" alt="Slide 193" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide194.png" alt="Slide 194" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide195.png" alt="Slide 195" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide196.png" alt="Slide 196" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide197.png" alt="Slide 197" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide198.png" alt="Slide 198" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide199.png" alt="Slide 199" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide200.png" alt="Slide 200" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide201.png" alt="Slide 201" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide202.png" alt="Slide 202" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide203.png" alt="Slide 203" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide204.png" alt="Slide 204" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide205.png" alt="Slide 205" />
+            </section>
+        <//>`;
+}
+
+export default Lesson07;
