+++ slides/src/Slides/Lessons/Lesson08.js
@@ -1,0 +1,112 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson08() {
+    return html`
+        <${Lesson} title="Functions" lessonId="lesson_08" subtitle="Lesson 8">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide207.png" alt="Slide 207" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide208.png" alt="Slide 208" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide209.png" alt="Slide 209" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide210.png" alt="Slide 210" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide211.png" alt="Slide 211" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide212.png" alt="Slide 212" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide213.png" alt="Slide 213" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide214.png" alt="Slide 214" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide215.png" alt="Slide 215" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide216.png" alt="Slide 216" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide217.png" alt="Slide 217" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide218.png" alt="Slide 218" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide219.png" alt="Slide 219" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide220.png" alt="Slide 220" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide221.png" alt="Slide 221" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide222.png" alt="Slide 222" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide223.png" alt="Slide 223" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide224.png" alt="Slide 224" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide225.png" alt="Slide 225" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide226.png" alt="Slide 226" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide227.png" alt="Slide 227" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide228.png" alt="Slide 228" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide229.png" alt="Slide 229" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide230.png" alt="Slide 230" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide231.png" alt="Slide 231" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide232.png" alt="Slide 232" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide233.png" alt="Slide 233" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide234.png" alt="Slide 234" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide235.png" alt="Slide 235" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide236.png" alt="Slide 236" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide237.png" alt="Slide 237" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide238.png" alt="Slide 238" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide239.png" alt="Slide 239" />
+            </section>
+        <//>`;
+}
+
+export default Lesson08;
