+++ slides/src/Slides/Lessons/Lesson09.js
@@ -1,0 +1,169 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson09() {
+    return html`
+        <${Lesson} title="Data Types" lessonId="lesson_09" subtitle="Lesson 9">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide241.png" alt="Slide 241" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide242.png" alt="Slide 242" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide243.png" alt="Slide 243" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide244.png" alt="Slide 244" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide245.png" alt="Slide 245" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide246.png" alt="Slide 246" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide247.png" alt="Slide 247" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide248.png" alt="Slide 248" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide249.png" alt="Slide 249" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide250.png" alt="Slide 250" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide251.png" alt="Slide 251" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide252.png" alt="Slide 252" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide253.png" alt="Slide 253" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide254.png" alt="Slide 254" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide255.png" alt="Slide 255" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide256.png" alt="Slide 256" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide257.png" alt="Slide 257" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide258.png" alt="Slide 258" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide259.png" alt="Slide 259" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide260.png" alt="Slide 260" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide261.png" alt="Slide 261" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide262.png" alt="Slide 262" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide263.png" alt="Slide 263" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide264.png" alt="Slide 264" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide265.png" alt="Slide 265" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide266.png" alt="Slide 266" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide267.png" alt="Slide 267" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide268.png" alt="Slide 268" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide269.png" alt="Slide 269" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide270.png" alt="Slide 270" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide271.png" alt="Slide 271" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide272.png" alt="Slide 272" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide273.png" alt="Slide 273" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide274.png" alt="Slide 274" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide275.png" alt="Slide 275" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide276.png" alt="Slide 276" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide277.png" alt="Slide 277" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide278.png" alt="Slide 278" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide279.png" alt="Slide 279" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide280.png" alt="Slide 280" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide281.png" alt="Slide 281" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide282.png" alt="Slide 282" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide283.png" alt="Slide 283" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide284.png" alt="Slide 284" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide285.png" alt="Slide 285" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide286.png" alt="Slide 286" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide287.png" alt="Slide 287" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide288.png" alt="Slide 288" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide289.png" alt="Slide 289" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide290.png" alt="Slide 290" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide291.png" alt="Slide 291" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide292.png" alt="Slide 292" />
+            </section>
+        <//>`;
+}
+
+export default Lesson09;
