+++ slides/src/Slides/Lessons/Lesson10.js
@@ -1,0 +1,88 @@
+import htm from "htm";
+import { createElement } from "react";
+import { Lesson } from "../Layouts/index.js";
+
+const html = htm.bind(createElement);
+
+function Lesson10() {
+    return html`
+        <${Lesson} title="Libraries" lessonId="lesson_10" subtitle="Lesson 10">
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide294.png" alt="Slide 294" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide295.png" alt="Slide 295" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide296.png" alt="Slide 296" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide297.png" alt="Slide 297" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide298.png" alt="Slide 298" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide299.png" alt="Slide 299" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide300.png" alt="Slide 300" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide301.png" alt="Slide 301" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide302.png" alt="Slide 302" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide303.png" alt="Slide 303" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide304.png" alt="Slide 304" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide305.png" alt="Slide 305" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide306.png" alt="Slide 306" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide307.png" alt="Slide 307" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide308.png" alt="Slide 308" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide309.png" alt="Slide 309" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide310.png" alt="Slide 310" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide311.png" alt="Slide 311" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide312.png" alt="Slide 312" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide313.png" alt="Slide 313" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide314.png" alt="Slide 314" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide315.png" alt="Slide 315" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide316.png" alt="Slide 316" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide317.png" alt="Slide 317" />
+            </section>
+            <section>
+                <img class="r-stretch" src="images/24q4-slides/Slide318.png" alt="Slide 318" />
+            </section>
+        <//>`;
+}
+
+export default Lesson10;
