+++ slides/src/Slides/Slides.js
@@ -1,7 +1,17 @@
-
 import htm from "htm";
 import { createElement } from "react";
-import TitleSlide from "./TitleSlide.js";
+import TitleSlide from "./Layouts/TitleSlide.js";
+import Lesson00 from "./Lessons/Lesson00.js";
+import Lesson01 from "./Lessons/Lesson01.js";
+import Lesson02 from "./Lessons/Lesson02.js";
+import Lesson03 from "./Lessons/Lesson03.js";
+import Lesson04 from "./Lessons/Lesson04.js";
+import Lesson05 from "./Lessons/Lesson05.js";
+import Lesson06 from "./Lessons/Lesson06.js";
+import Lesson07 from "./Lessons/Lesson07.js";
+import Lesson08 from "./Lessons/Lesson08.js";
+import Lesson09 from "./Lessons/Lesson09.js";
+import Lesson10 from "./Lessons/Lesson10.js";
 
 const html = htm.bind(createElement);
 
@@ -14,957 +24,17 @@ function Slides() {
         <section>
             <img class="r-stretch" src="images/24q4-slides/Slide2.png" alt="Slide 2" />
         </section>
-        <${TitleSlide} title="Git Fundamentals" slideId="lesson_00" subtitle="Lesson 0" />
-        <section class="ml-bullet-slide">
-            <h3>What we're doing</h3>
-            <ul>
-                <li class="fragment">Why we need version control</li>
-                <li class="fragment">What is Git and GitHub?</li>
-                <li class="fragment">Working with forks and branches</li>
-                <li class="fragment">Learning the basic workflow</li>
-            </ul>
-        </section>
-        <section>
-            <p>Why we need <em>version control</em></p>
-        </section>
-        <section class="ml-bullet-slide">
-            <h3>Why we need <em>version control</em></h3>
-            <ul>
-                <li class="fragment">Track changes over time</li>
-                <li class="fragment">Collaborate with others</li>
-                <li class="fragment">Revert to previous versions</li>
-                <li class="fragment">Backup your work</li>
-            </ul>
-        </section>
-        <section>
-            <p>What is <em>Git</em> and <em>GitHub</em>?</p>
-        </section>
-        <section>
-            <p><em>Git</em> is a distributed version control system (VCS) that works well for branching and is fast</p>
-        </section>
-        <section>
-            <p><em>GitHub</em> is a platform for hosting Git repositories and features tools for managing software collaboration</p>
-        </section>
-        <section>
-            <p>Working with <em>Forks</em> and <em>Branches</em></p>
-        </section>
-        <section>
-            <p>A <em>branch</em> is a copy of the repo that allows you to make changes and merge them back later</p>
-        </section>
-        <section>
-            <p><em>Forks</em> are just branched copies of other user's repos that you own and modify</p>
-        </section>
-        <section class="ml-bullet-slide">
-            <h3>Why we <em>branch</em></h3>
-            <ul>
-                <li class="fragment">Work on features or fixes without impacting work happening in other places</li>
-                <li class="fragment">Run tests or deployments</li>
-                <li class="fragment">Experiment with changes safely</li>
-            </ul>
-        </section>
-        <section>
-            <p>Learning the <em>basic workflow</em></p>
-        </section>
-        <section class="ml-bullet-slide">
-            <h3>The <em>basic workflow</em></h3>
-            <ol>
-                <li class="fragment">Pull the latest changes from other devs</li>
-                <li class="fragment">Change some more stuff</li>
-                <li class="fragment">Commit to git locally</li>
-                <li class="fragment">Push to server remotely (GitHub)</li>
-                <li class="fragment">Repeat 1–4 a bunch</li>
-                <li class="fragment">Create a pull request (PR) when done</li>
-            </ol>
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide16.png" alt="Slide 16" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide17.png" alt="Slide 17" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide18.png" alt="Slide 18" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide19.png" alt="Slide 19" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide20.png" alt="Slide 20" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide21.png" alt="Slide 21" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide22.png" alt="Slide 22" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide23.png" alt="Slide 23" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide24.png" alt="Slide 24" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide25.png" alt="Slide 25" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide26.png" alt="Slide 26" />
-        </section>
-        <${TitleSlide} title="Getting to Know the IDE" slideId="lesson_01" subtitle="Lesson 1" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide28.png" alt="Slide 28" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide29.png" alt="Slide 29" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide30.png" alt="Slide 30" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide31.png" alt="Slide 31" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide32.png" alt="Slide 32" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide33.png" alt="Slide 33" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide34.png" alt="Slide 34" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide35.png" alt="Slide 35" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide36.png" alt="Slide 36" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide37.png" alt="Slide 37" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide38.png" alt="Slide 38" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide39.png" alt="Slide 39" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide40.png" alt="Slide 40" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide41.png" alt="Slide 41" />
-        </section>
-        <${TitleSlide} title="Exploring the Terminal" slideId="lesson_02" subtitle="Lesson 2" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide43.png" alt="Slide 43" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide44.png" alt="Slide 44" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide45.png" alt="Slide 45" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide46.png" alt="Slide 46" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide47.png" alt="Slide 47" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide48.png" alt="Slide 48" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide49.png" alt="Slide 49" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide50.png" alt="Slide 50" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide51.png" alt="Slide 51" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide52.png" alt="Slide 52" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide53.png" alt="Slide 53" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide54.png" alt="Slide 54" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide55.png" alt="Slide 55" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide56.png" alt="Slide 56" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide57.png" alt="Slide 57" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide58.png" alt="Slide 58" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide59.png" alt="Slide 59" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide60.png" alt="Slide 60" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide61.png" alt="Slide 61" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide62.png" alt="Slide 62" />
-        </section>
-        <${TitleSlide} title="How Computers Work" slideId="lesson_03" subtitle="Lesson 3" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide64.png" alt="Slide 64" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide65.png" alt="Slide 65" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide66.png" alt="Slide 66" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide67.png" alt="Slide 67" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide68.png" alt="Slide 68" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide69.png" alt="Slide 69" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide70.png" alt="Slide 70" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide71.png" alt="Slide 71" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide72.png" alt="Slide 72" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide73.png" alt="Slide 73" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide74.png" alt="Slide 74" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide75.png" alt="Slide 75" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide76.png" alt="Slide 76" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide77.png" alt="Slide 77" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide78.png" alt="Slide 78" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide79.png" alt="Slide 79" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide80.png" alt="Slide 80" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide81.png" alt="Slide 81" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide82.png" alt="Slide 82" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide83.png" alt="Slide 83" />
-        </section>
-        <${TitleSlide} title="How to be a Great SWE" slideId="lesson_04" subtitle="Lesson 4" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide85.png" alt="Slide 85" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide86.png" alt="Slide 86" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide87.png" alt="Slide 87" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide88.png" alt="Slide 88" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide89.png" alt="Slide 89" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide90.png" alt="Slide 90" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide91.png" alt="Slide 91" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide92.png" alt="Slide 92" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide93.png" alt="Slide 93" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide94.png" alt="Slide 94" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide95.png" alt="Slide 95" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide96.png" alt="Slide 96" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide97.png" alt="Slide 97" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide98.png" alt="Slide 98" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide99.png" alt="Slide 99" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide100.png" alt="Slide 100" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide101.png" alt="Slide 101" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide102.png" alt="Slide 102" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide103.png" alt="Slide 103" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide104.png" alt="Slide 104" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide105.png" alt="Slide 105" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide106.png" alt="Slide 106" />
-        </section>
-        <${TitleSlide} title="Software Dev Life Cycle" slideId="lesson_05" subtitle="Lesson 5" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide108.png" alt="Slide 108" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide109.png" alt="Slide 109" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide110.png" alt="Slide 110" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide111.png" alt="Slide 111" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide112.png" alt="Slide 112" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide113.png" alt="Slide 113" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide114.png" alt="Slide 114" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide115.png" alt="Slide 115" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide116.png" alt="Slide 116" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide117.png" alt="Slide 117" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide118.png" alt="Slide 118" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide119.png" alt="Slide 119" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide120.png" alt="Slide 120" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide121.png" alt="Slide 121" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide122.png" alt="Slide 122" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide123.png" alt="Slide 123" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide124.png" alt="Slide 124" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide125.png" alt="Slide 125" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide126.png" alt="Slide 126" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide127.png" alt="Slide 127" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide128.png" alt="Slide 128" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide129.png" alt="Slide 129" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide130.png" alt="Slide 130" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide131.png" alt="Slide 131" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide132.png" alt="Slide 132" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide133.png" alt="Slide 133" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide134.png" alt="Slide 134" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide135.png" alt="Slide 135" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide136.png" alt="Slide 136" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide137.png" alt="Slide 137" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide138.png" alt="Slide 138" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide139.png" alt="Slide 139" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide140.png" alt="Slide 140" />
-        </section>
-        <${TitleSlide} title="Statements and Variables" slideId="lesson_06" subtitle="Lesson 6" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide142.png" alt="Slide 142" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide143.png" alt="Slide 143" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide144.png" alt="Slide 144" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide145.png" alt="Slide 145" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide146.png" alt="Slide 146" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide147.png" alt="Slide 147" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide148.png" alt="Slide 148" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide149.png" alt="Slide 149" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide150.png" alt="Slide 150" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide151.png" alt="Slide 151" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide152.png" alt="Slide 152" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide153.png" alt="Slide 153" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide154.png" alt="Slide 154" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide155.png" alt="Slide 155" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide156.png" alt="Slide 156" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide157.png" alt="Slide 157" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide158.png" alt="Slide 158" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide159.png" alt="Slide 159" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide160.png" alt="Slide 160" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide161.png" alt="Slide 161" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide162.png" alt="Slide 162" />
-        </section>
-        <${TitleSlide} title="Control Flow" slideId="lesson_07" subtitle="Lesson 7" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide164.png" alt="Slide 164" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide165.png" alt="Slide 165" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide166.png" alt="Slide 166" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide167.png" alt="Slide 167" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide168.png" alt="Slide 168" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide169.png" alt="Slide 169" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide170.png" alt="Slide 170" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide171.png" alt="Slide 171" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide172.png" alt="Slide 172" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide173.png" alt="Slide 173" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide174.png" alt="Slide 174" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide175.png" alt="Slide 175" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide176.png" alt="Slide 176" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide177.png" alt="Slide 177" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide178.png" alt="Slide 178" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide179.png" alt="Slide 179" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide180.png" alt="Slide 180" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide181.png" alt="Slide 181" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide182.png" alt="Slide 182" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide183.png" alt="Slide 183" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide184.png" alt="Slide 184" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide185.png" alt="Slide 185" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide186.png" alt="Slide 186" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide187.png" alt="Slide 187" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide188.png" alt="Slide 188" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide189.png" alt="Slide 189" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide190.png" alt="Slide 190" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide191.png" alt="Slide 191" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide192.png" alt="Slide 192" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide193.png" alt="Slide 193" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide194.png" alt="Slide 194" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide195.png" alt="Slide 195" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide196.png" alt="Slide 196" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide197.png" alt="Slide 197" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide198.png" alt="Slide 198" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide199.png" alt="Slide 199" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide200.png" alt="Slide 200" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide201.png" alt="Slide 201" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide202.png" alt="Slide 202" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide203.png" alt="Slide 203" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide204.png" alt="Slide 204" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide205.png" alt="Slide 205" />
-        </section>
-        <${TitleSlide} title="Functions" slideId="lesson_08" subtitle="Lesson 8" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide207.png" alt="Slide 207" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide208.png" alt="Slide 208" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide209.png" alt="Slide 209" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide210.png" alt="Slide 210" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide211.png" alt="Slide 211" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide212.png" alt="Slide 212" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide213.png" alt="Slide 213" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide214.png" alt="Slide 214" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide215.png" alt="Slide 215" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide216.png" alt="Slide 216" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide217.png" alt="Slide 217" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide218.png" alt="Slide 218" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide219.png" alt="Slide 219" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide220.png" alt="Slide 220" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide221.png" alt="Slide 221" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide222.png" alt="Slide 222" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide223.png" alt="Slide 223" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide224.png" alt="Slide 224" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide225.png" alt="Slide 225" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide226.png" alt="Slide 226" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide227.png" alt="Slide 227" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide228.png" alt="Slide 228" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide229.png" alt="Slide 229" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide230.png" alt="Slide 230" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide231.png" alt="Slide 231" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide232.png" alt="Slide 232" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide233.png" alt="Slide 233" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide234.png" alt="Slide 234" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide235.png" alt="Slide 235" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide236.png" alt="Slide 236" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide237.png" alt="Slide 237" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide238.png" alt="Slide 238" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide239.png" alt="Slide 239" />
-        </section>
-        <${TitleSlide} title="Data Types" slideId="lesson_09" subtitle="Lesson 9" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide241.png" alt="Slide 241" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide242.png" alt="Slide 242" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide243.png" alt="Slide 243" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide244.png" alt="Slide 244" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide245.png" alt="Slide 245" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide246.png" alt="Slide 246" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide247.png" alt="Slide 247" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide248.png" alt="Slide 248" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide249.png" alt="Slide 249" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide250.png" alt="Slide 250" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide251.png" alt="Slide 251" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide252.png" alt="Slide 252" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide253.png" alt="Slide 253" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide254.png" alt="Slide 254" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide255.png" alt="Slide 255" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide256.png" alt="Slide 256" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide257.png" alt="Slide 257" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide258.png" alt="Slide 258" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide259.png" alt="Slide 259" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide260.png" alt="Slide 260" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide261.png" alt="Slide 261" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide262.png" alt="Slide 262" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide263.png" alt="Slide 263" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide264.png" alt="Slide 264" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide265.png" alt="Slide 265" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide266.png" alt="Slide 266" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide267.png" alt="Slide 267" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide268.png" alt="Slide 268" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide269.png" alt="Slide 269" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide270.png" alt="Slide 270" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide271.png" alt="Slide 271" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide272.png" alt="Slide 272" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide273.png" alt="Slide 273" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide274.png" alt="Slide 274" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide275.png" alt="Slide 275" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide276.png" alt="Slide 276" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide277.png" alt="Slide 277" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide278.png" alt="Slide 278" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide279.png" alt="Slide 279" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide280.png" alt="Slide 280" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide281.png" alt="Slide 281" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide282.png" alt="Slide 282" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide283.png" alt="Slide 283" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide284.png" alt="Slide 284" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide285.png" alt="Slide 285" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide286.png" alt="Slide 286" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide287.png" alt="Slide 287" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide288.png" alt="Slide 288" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide289.png" alt="Slide 289" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide290.png" alt="Slide 290" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide291.png" alt="Slide 291" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide292.png" alt="Slide 292" />
-        </section>
-        <${TitleSlide} title="Libraries" slideId="lesson_10" subtitle="Lesson 10" />
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide294.png" alt="Slide 294" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide295.png" alt="Slide 295" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide296.png" alt="Slide 296" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide297.png" alt="Slide 297" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide298.png" alt="Slide 298" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide299.png" alt="Slide 299" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide300.png" alt="Slide 300" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide301.png" alt="Slide 301" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide302.png" alt="Slide 302" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide303.png" alt="Slide 303" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide304.png" alt="Slide 304" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide305.png" alt="Slide 305" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide306.png" alt="Slide 306" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide307.png" alt="Slide 307" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide308.png" alt="Slide 308" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide309.png" alt="Slide 309" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide310.png" alt="Slide 310" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide311.png" alt="Slide 311" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide312.png" alt="Slide 312" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide313.png" alt="Slide 313" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide314.png" alt="Slide 314" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide315.png" alt="Slide 315" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide316.png" alt="Slide 316" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide317.png" alt="Slide 317" />
-        </section>
-        <section>
-            <img class="r-stretch" src="images/24q4-slides/Slide318.png" alt="Slide 318" />
-        </section>
+        <${Lesson00} />
+        <${Lesson01} />
+        <${Lesson02} />
+        <${Lesson03} />
+        <${Lesson04} />
+        <${Lesson05} />
+        <${Lesson06} />
+        <${Lesson07} />
+        <${Lesson08} />
+        <${Lesson09} />
+        <${Lesson10} />
         <${TitleSlide} title="Data Structures: Strings & Arrays" slideId="lesson_11" subtitle="Lesson 11" />
         <section>
             <img class="r-stretch" src="images/24q4-slides/Slide320.png" alt="Slide 320" />
